log changes to build token
