+++ lib/requestHandlers.js
@@ -3,6 +3,7 @@
 const auth = require('./auth'),
       schemas  = require('../models/models'),
       helpers  = require('../models/helpers'),
+      saveLog  = require('./saveLog').save,
       documentToObject = require('./documentToObject');
 
 var m = {},
@@ -439,6 +440,12 @@ e.rollBuildToken = (req, res) => {
                                         if (err) {
                                             sendError(500, `Unable to update buildToken due to interal error on Shipment ${shipment} Environment ${environment}`, err, msg);
                                         } else {
+                                            saveLog({buildToken: '**previous**'}, {buildToken: '**current**'}, {
+                                                shipment:    shipment,
+                                                environment: environment,
+                                                username:    req.body.username
+                                            });
+
                                             res.json(documentToObject(saveResult));
                                         }
                                     });
